requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
streamlit==1.38.0
pydantic==2.9.2
requests==2.32.3
python-multipart==0.0.9

# Additional Requirements
python-dotenv==1.0.1
# Local Dependencies and Requirements not to be used in production